Add operator check and update item features
The command executor now requires player to be an Operator for execution. A count feature is added to allow giving more than one ExpandBorderItem to the player at a time. The ExpandBorderItem presentation is modified with a change in Material to GOAT_HORN and an additional lore text. Updates are also made on ShopItem with additions of comments to variables and a new 'allowList' variable.